+++ lib/omni_service/parallel.rb
@@ -29,7 +29,7 @@ class OmniService::Parallel
   include OmniService::Inspect.new(:components)
   include OmniService::Strict
 
-  param :components, OmniService::Types::Array.of(OmniService::Types::Interface(:call)).constrained(min_size: 1)
+  param :components, OmniService::Types::Array.of(OmniService::Types::Callable).constrained(min_size: 1)
   option :pack_params, OmniService::Types::Bool, default: proc { false }
 
   def initialize(*args, **)